payout amount can accept float amount

+++ database/migrations/2023_04_30_103930_create_assets_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('asset_type');
             $table->string('asset_name');
             $table->string('reward_level');
-            $table->integer('payout_amount');
+            $table->decimal('payout_amount', 10, 2);
             $table->timestamps();
         });
     }
